+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to the Ajrly Platform will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.17] - 2025-11-23
+
+**Author:** Nasr Aldin
+
+### fix
+
+fix: sonar 101
+
+### File Changed
+
+- CHANGELOG.md
+- package.json
+
+---
+
+## [0.3.16] - 2025-11-23
+
+**Author:** Nasr Aldin
+
+### fix
+
+fix: sonar 101
+
+### File Changed
+
+- package.json
+- sonar-project.properties
+- vitest.config.ts
+
+---
+
 ## [0.3.13] - 2025-11-23
 
 **Author:** Nasr Aldin